pyproject.toml
[project]
name = "pbakelompok1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "newspaper4k>=0.9.5",
    "nltk>=3.9.4",
    "pandas>=3.0.2",
    "playwright>=1.58.0",
    "requests>=2.33.1",
    "tqdm>=4.67.3",
]
